+++ examples/storage_pool.rb
@@ -15,7 +15,7 @@ my_client = {
   url: ENV['ONEVIEWSDK_URL'],
   user: ENV['ONEVIEWSDK_USER'],
   password: ENV['ONEVIEWSDK_PASSWORD'],
-  api_version: 2000
+  api_version: 2200
 }
 
 # Ensures the Storage Pool is managed by the HPE OneView appliance