+++ examples/logical_interconnect.rb
@@ -19,7 +19,7 @@ my_client = {
   user: ENV['ONEVIEWSDK_USER'],
   password: ENV['ONEVIEWSDK_PASSWORD'],
   api_variant: 'Synergy',
-  api_version: 2000
+  api_version: 2200
 }
 
 # Variable declaration