+++ examples/image_streamer/artifact_bundle.rb
@@ -21,7 +21,7 @@ oneview_client = {
 i3s_client = {
   url: ENV['I3S_URL'],
   oneview_client: oneview_client,
-  api_version: 1600
+  api_version: 2000
 }
 
 # Creates the artifact bundle and assigns resources to it.