+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 ### Oneview Features supported
 - oneview_connection_template
+- oneview_enclosure
+- oneview_enclosure_group
 - oneview_ethernet_network
 - oneview_fc_network
 - oneview_fcoe_network
 - oneview_network_set
 - oneview_scope
+- oneview_server_hardware
+- oneview_server_hardware_type
 - oneview_uplink_set
 
 ## 3.6.0