+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - oneview_ethernet_network
 - oneview_fc_network
 - oneview_fcoe_network
+- oneview_interconnect
+- oneview_logical_enclosure
+- oneview_logical_interconnect
+- oneview_logical_interconnect_group
 - oneview_network_set
 - oneview_scope
 - oneview_uplink_set