+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Oneview Features supported
 - oneview_connection_template
+- oneview_enclosure
+- oneview_enclosure_group
 - oneview_ethernet_network
 - oneview_fc_network
 - oneview_fcoe_network
@@ -12,6 +14,8 @@
 - oneview_network_set
 - oneview_scope
 - oneview_server_certificate
+- oneview_server_hardware
+- oneview_server_hardware_type
 - oneview_uplink_set
 
 ## 3.6.0